requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==5.1.1
frozenlist==1.4.1
fsspec==2024.9.0
gcsfs==2024.9.0.post1
google-api-core==2.19.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.9
multidict==6.1.0
numpy==2.1.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tzdata==2024.1
urllib3==2.2.3
yarl==1.11.1
